# api/openapi/management-v1.yaml
openapi: 3.1.0
info:
  title: MarketplaceServer Management API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Authoritative wire contract for currently implemented management operations
    under `/api/v1`.

    This document contains deployed login, health, current-user personal access
    token, Plugin lifecycle, and Plugin Version operations. Future Marketplace,
    team, and distribution credential management paths remain in
    `management-v1-design.yaml` and are not evidence of runtime availability.

    Git Smart HTTP, distribution endpoints, frontend routes, and the external
    Claude Code `marketplace.json` schema are outside this contract.
  x-document-status: approved
  x-implementation-conformance: implemented
...
servers:
  - url: /
    description: Same-origin MarketplaceServer deployment
...
paths:
...
  /api/v1/me/tokens:
    get:
      operationId: listPersonalAccessTokens
      tags: [Personal access tokens]
      summary: List the current user's personal access tokens
      description: |
        Returns an exact owner-scoped page ordered by creation time and ID,
        newest first. Metadata includes active, expired, and revoked tokens but
        never includes plaintext or HMAC secret material.
      security:
        - BearerJWT: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Exact owner-scoped token page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPageSuccess"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createPersonalAccessToken
      tags: [Personal access tokens]
      summary: Create a personal access token
      description: |
        Creates an owner-scoped PAT using one cumulative capability preset.
        Unknown JSON properties are currently ignored. The repeatably revealable
        plaintext is returned in this response and is protected from caching.
      security:
        - BearerJWT: []
      requestBody:
        required: true
        description: |
          The handler decodes JSON but does not currently enforce the Content-Type
          header. Request bodies are limited to 65,536 bytes.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: Token created with owner-revealable plaintext.
          headers:
            Cache-Control:
              $ref: "#/components/headers/SecretCacheControl"
            Pragma:
              $ref: "#/components/headers/SecretPragma"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedTokenSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/tokens/{tokenId}:
    parameters:
      - $ref: "#/components/parameters/TokenId"
    delete:
      operationId: revokePersonalAccessToken
      tags: [Personal access tokens]
      summary: Revoke an owner-scoped personal access token
      description: |
        Revocation is idempotent. A malformed, noncanonical, missing, or
        other-owner token identifier is intentionally hidden behind the same 404.
      security:
        - BearerJWT: []
      responses:
        "204":
          description: Token revoked idempotently; response body is empty.
        "401":
          $ref: "#/components/responses/BearerUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/TokenNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/tokens/{tokenId}/reveal:
    parameters:
      - $ref: "#/components/parameters/TokenId"
    post:
      operationId: revealPersonalAccessToken
      tags: [Personal access tokens]
      summary: Reveal an owner-scoped personal access token
      description: |
        Requires both a management Bearer JWT and the current account password.
        The owner may reveal active, expired, or revoked PAT plaintext. Unknown
        JSON properties are currently ignored and the response is not cacheable.
      security:
        - BearerJWT: []
      requestBody:
        required: true
        description: |
          The handler decodes JSON but does not currently enforce the Content-Type
          header. Request bodies are limited to 65,536 bytes.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevealTokenRequest"
      responses:
        "200":
          description: Token plaintext and current metadata were returned to its owner.
          headers:
            Cache-Control:
              $ref: "#/components/headers/SecretCacheControl"
            Pragma:
              $ref: "#/components/headers/SecretPragma"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevealedTokenSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/RevealUnauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/TokenNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Stateless management JWT signed with HS256. Issued tokens contain exactly
        canonical username, iat, and exp claims and have a fixed 30-day lifetime.
        Account passwords and PATs are not accepted as management Bearer tokens.
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: One-based page number; defaults to 1.
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      description: Page size; defaults to 20 and must not exceed 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    TokenId:
      name: tokenId
      in: path
      required: true
      description: Canonical lowercase hyphenated UUID identifying the PAT.
      schema:
        $ref: "#/components/schemas/UUID"
...
  headers:
    RequestId:
      description: Framework-generated ULID matching the error body's requestId.
      schema:
        type: string
        pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    BearerChallenge:
      description: Management Bearer authentication challenge.
      schema:
        type: string
        const: 'Bearer realm="MarketplaceServer Management"'
    SecretCacheControl:
      description: Prevents storage of a secret-bearing response.
      schema:
        type: string
        const: private, no-store
    SecretPragma:
      description: Compatibility cache prevention for a secret-bearing response.
      schema:
        type: string
        const: no-cache
  responses:
    BadRequest:
      description: The body is missing, unreadable, malformed, oversized, top-level null, or contains a trailing JSON value.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: bad_request
            message: request body is invalid
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
...
    BearerUnauthenticated:
      description: The management Bearer JWT is missing or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          $ref: "#/components/headers/BearerChallenge"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: authentication is required
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
    RevealUnauthenticated:
      description: |
        The management Bearer JWT is missing or invalid, or the password
        confirmation failed. Bearer-authentication failures include
        `WWW-Authenticate`; password-confirmation failures omit it. The response
        header is therefore optional.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          $ref: "#/components/headers/BearerChallenge"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated principal is not authorized for this action and resource.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: forbidden
            message: permission denied
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
    TokenNotFound:
      description: The token is malformed, missing, noncanonical, or not owned by the current user.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: not_found
            message: token not found
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
...
    ValidationFailed:
      description: Query syntax or token input violates the documented constraints.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: validation_failed
            message: request is invalid
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
    InternalError:
      description: Internal failure without implementation detail disclosure.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: internal_error
            message: internal server error
            requestId: 01J51X8R6M4B9D5VYN2Q3H7KCT
  schemas:
    UUID:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: Canonical lowercase hyphenated UUID text.
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp.
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
          pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
        details:
          type: string
          description: Optional safe text; clients must branch on code, not this value.
...
    TokenPreset:
      type: string
      enum: [sub-read, git-clone, git-write]
      description: Cumulative subscription/Git capability; never broader than the owner's current authorization.
    TokenStatus:
      type: string
      enum: [active, expired, revoked]
    TokenMetadata:
      type: object
      additionalProperties: false
      required: [id, name, preset, status, expiresAt, lastUsedAt, revokedAt, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
          minLength: 1
          description: Trimmed human-readable name, limited by the runtime to 128 UTF-8 bytes.
        preset:
          $ref: "#/components/schemas/TokenPreset"
        status:
          $ref: "#/components/schemas/TokenStatus"
        expiresAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        lastUsedAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        revokedAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    CreateTokenRequest:
      type: object
      required: [name, preset]
      properties:
        name:
          type: string
          minLength: 1
          description: Trimmed value must be nonempty and at most 128 UTF-8 bytes.
        preset:
          $ref: "#/components/schemas/TokenPreset"
        expiresAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
          description: Optional expiration strictly later than creation time.
    RevealTokenRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          writeOnly: true
    RevealedToken:
      type: object
      additionalProperties: false
      required: [id, name, preset, status, expiresAt, lastUsedAt, revokedAt, createdAt, token]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
          minLength: 1
          description: Trimmed human-readable name, limited by the runtime to 128 UTF-8 bytes.
        preset:
          $ref: "#/components/schemas/TokenPreset"
        status:
          $ref: "#/components/schemas/TokenStatus"
        expiresAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        lastUsedAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        revokedAt:
          oneOf:
            - $ref: "#/components/schemas/Timestamp"
            - type: "null"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        token:
          type: string
          minLength: 48
          maxLength: 48
          pattern: "^mpsk_[A-Za-z0-9_-]{43}$"
          description: Repeatably revealable plaintext PAT.
    CreatedTokenSuccess:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RevealedToken"
    RevealedTokenSuccess:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RevealedToken"
...
    TokenPageData:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TokenMetadata"
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
...
    TokenPageSuccess:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/TokenPageData"
...